refactor: productionize CLI architecture
