+++ src/lib/fetchConfig.ts
@@ -1,3 +1,15 @@
+import { config } from "./config";
+
 export function getFetchOptions(additionalOptions: RequestInit = {}): RequestInit {
-  return additionalOptions;
+  if (config.sslVerify) return additionalOptions;
+
+  const bunOptions = additionalOptions as RequestInit & { tls?: { rejectUnauthorized?: boolean } };
+
+  return {
+    ...additionalOptions,
+    tls: {
+      rejectUnauthorized: false,
+      ...bunOptions.tls,
+    },
+  } as RequestInit;
 }
